Making BlockDevice take &mut self
Reading and Writing from/to a block device cannot reasonably be done constly.
A BlockDevice will almost always need some form of mutability to be read from/written to
(ex: std::io::File::Read is &mut self to be able to move the cursor,
CachedBlockDevice needs to be able to modify its cache, etc.).

Imposing interior mutability makes deriving BlockDevice an hassle, where the concurrency
should in fact be guaranteed by the holder of BlockDevice, not by the BlockDevice itself.