Fix std feature dependencies
Having the impl BlockDevice for std::fs::File behind a #[cfg(test)] is not satisfying,
as when built as a dependency the cfg(test) is not propagated to us.

We need to put it behind a feature.

Unfortunately for us, cargo's feature management is pure masochism.
It seems there is no way to:

* Enable a specific set of features when building for test.
* Turn 'off' a transitive feature if we choose to declare it on by default.
* Declare some features as mutually exclusive.
* Choose whether to enable a transitive feature based on the combination of two features.

The solution (and I'm really not proud of it) was to declare two features for the same
conditional compilation unit, one for nightly and one for stable.

A crate that depends on storage_device should be written in the following way:

```
[package]
name = "my_awesome_crate"
...

[dependencies.storage_device]
verion = "..."
default-features = false

[features]
 # Default behaviour for my_awesome_crate is to be built with std.
default = ["std"]
 # Link my_awesome_crate with std; used for unit tests.
 #
 # Transitively enables storage_device's `std` feature.
std = ["storage_device/std"]
 # Enable cache-block-device but in nightly. Used for no_std compilation.
 #
 # Transitively enables storage_device's `cached-block-device-nightly` feature,
 # which in turns internally enables lru's `nightly` feature.
 #
 # hashbrown just stands as an example for other crates you might want to have features enabled for.
nightly = ["hashbrown/nightly", "storage_device/cached-block-device-nightly"]

```

Also PR to lru was mainlined, so we don't have to use the fork anymore.