removed struct StorageBlockDevice<B>
StorageBlockDevice<B> is just an unnecessary wrapper around a BlockDevice 'B'.
We can just derive StorageDevice for every 'B'.

This means that every BlockDevice is also a StorageDevice through either the
blanket implementation, or some user-defined specialised implementation.

Having to make the user add a wrapper around its BlockDevice is a real pain.